… pkg/tsdb/prometheus/querydata/exemplars.go
… pkg/util/converter/testdata/prom-exemplars-b-frame.jsonc
… pkg/util/converter/testdata/prom-exemplars-b-wide-frame.jsonc