<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

-->

**What this PR does / why we need it**:

```
benchstat old.txt new.txt
name             old time/op    new time/op    delta
ExemplarJson-32    25.4ms ± 7%    11.0ms ± 9%  -56.44%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
ExemplarJson-32    9.58MB ± 0%    5.17MB ± 0%  -46.01%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
ExemplarJson-32      116k ± 0%       45k ± 0%  -61.41%  (p=0.000 n=10+10)
```

**Which issue(s) this PR fixes**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer**: