@toddtreece 2022-06-06
closing this one for now. it should be easy to port the changes from https://github.com/grafana/grafana/pull/49858 in the future if we'd like to go this route